requirements.txt
django-cms<3.3
django-treebeard>=2.0

https://github.com/divio/djangocms-column/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-googlemap/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-inherit/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-link/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-style/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-file/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-picture/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-teaser/archive/master.zip?1454083311.69
https://github.com/divio/djangocms-video/archive/master.zip?1454083311.69
djangocms-admin-style>=1.0.6
djangocms-text-ckeditor>=2.8.0
Django<1.9
pytz
django-reversion>=1.8.7,<1.9
django-classy-tags>=0.3.4.1
html5lib
Pillow>=2.3
django-sekizai>=0.7
six
PyYAML==3.11
django-bower==5.0.4
django-happenings==0.3.3
MySQL-python==1.2.5
djangorestframework==3.3.2
django-newsletter==0.6
unicodecsv==0.14.1
django-extensions==1.6.1
sorl-thumbnail==12.4a1
Unidecode==0.04.19
surlex==0.2.0
cmsplugin-filer==1.0.1
django-filer==1.1.1
aldryn-apphooks-config==0.2.6
django-parler==1.6.1
django-taggit==0.18.0
django-taggit-autosuggest==0.2.8
django-taggit-templatetags==0.2.5
django-meta==0.3.2
django-meta-mixin==0.2.1
djangocms-blog==0.6.3
django-ajaximage==0.2.8
